useKeyboardnav is now tsx
